Update README.md
This project implements a Convolutional Neural Network (CNN) to classify waste images into predefined categories. The model aims to assist in efficient waste management by automating the sorting process. The results are visualized using a pie chart to show the distribution of waste types, either in the dataset or based on predictions.